avoid cast by using size_t counter variable

+++ server/scanjob.cpp
@@ -354,10 +354,10 @@ ScanJob::Private::applyGamma(std::vector<char>& ioData)
     uint16_t* s;
   } data = { ioData.data() };
   if (mGammaTable.size() == 1 << 8) {
-    for (int i = 0; i < (int)ioData.size(); ++i)
+    for (size_t i = 0; i < ioData.size(); ++i)
       data.b[i] = mGammaTable[data.b[i]];
   } else if (mGammaTable.size() == 1 << 16) {
-    for (int i = 0; i < (int)ioData.size() / 2; ++i)
+    for (size_t i = 0; i < ioData.size() / 2; ++i)
       data.s[i] = mGammaTable[data.s[i]];
   }
 }
